rename package name and class name
