+++ pom.xml
@@ -763,7 +763,7 @@
             </dependency>
 
             <dependency>
-                <groupId>com.yscope.presto</groupId>
+                <groupId>com.facebook.presto</groupId>
                 <artifactId>presto-clp</artifactId>
                 <version>${project.version}</version>
             </dependency>
