+++ presto-clp/src/main/java/com/facebook/presto/plugin/clp/ClpTableLayoutHandle.java
@@ -1,0 +1,73 @@
+/*
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.facebook.presto.plugin.clp;
+
+import com.facebook.presto.spi.ConnectorTableLayoutHandle;
+import com.fasterxml.jackson.annotation.JsonCreator;
+import com.fasterxml.jackson.annotation.JsonProperty;
+
+import java.util.Objects;
+import java.util.Optional;
+
+public class ClpTableLayoutHandle
+        implements ConnectorTableLayoutHandle
+{
+    private final ClpTableHandle table;
+    private final Optional<String> query;
+
+    @JsonCreator
+    public ClpTableLayoutHandle(@JsonProperty("table") ClpTableHandle table,
+                                @JsonProperty("query") Optional<String> query)
+    {
+        this.table = table;
+        this.query = query;
+    }
+
+    @JsonProperty
+    public ClpTableHandle getTable()
+    {
+        return table;
+    }
+
+    @JsonProperty
+    public Optional<String> getQuery()
+    {
+        return query;
+    }
+
+    @Override
+    public boolean equals(Object o)
+    {
+        if (this == o) {
+            return true;
+        }
+        if (o == null || getClass() != o.getClass()) {
+            return false;
+        }
+        ClpTableLayoutHandle that = (ClpTableLayoutHandle) o;
+        return Objects.equals(table, that.table);
+    }
+
+    @Override
+    public int hashCode()
+    {
+        return Objects.hash(table);
+    }
+
+    @Override
+    public String toString()
+    {
+        return table.toString();
+    }
+}
