--- presto-clp/src/main/java/com/yscope/presto/ClpConnector.java
@@ -1,90 +1,0 @@
-/*
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.yscope.presto;
-
-import com.facebook.airlift.bootstrap.LifeCycleManager;
-import com.facebook.airlift.log.Logger;
-import com.facebook.presto.spi.connector.Connector;
-import com.facebook.presto.spi.connector.ConnectorMetadata;
-import com.facebook.presto.spi.connector.ConnectorPlanOptimizerProvider;
-import com.facebook.presto.spi.connector.ConnectorSplitManager;
-import com.facebook.presto.spi.connector.ConnectorTransactionHandle;
-import com.facebook.presto.spi.function.FunctionMetadataManager;
-import com.facebook.presto.spi.function.StandardFunctionResolution;
-import com.facebook.presto.spi.transaction.IsolationLevel;
-
-import javax.inject.Inject;
-
-import static java.util.Objects.requireNonNull;
-
-public class ClpConnector
-        implements Connector
-{
-    private static final Logger log = Logger.get(ClpConnector.class);
-
-    private final LifeCycleManager lifeCycleManager;
-    private final ClpMetadata metadata;
-    private final ClpSplitManager splitManager;
-    private final FunctionMetadataManager functionManager;
-    private final StandardFunctionResolution functionResolution;
-
-    @Inject
-    public ClpConnector(LifeCycleManager lifeCycleManager,
-                        ClpMetadata metadata,
-                        ClpSplitManager splitManager,
-                        FunctionMetadataManager functionManager,
-                        StandardFunctionResolution functionResolution)
-    {
-        this.lifeCycleManager = requireNonNull(lifeCycleManager, "lifeCycleManager is null");
-        this.metadata = requireNonNull(metadata, "metadata is null");
-        this.splitManager = requireNonNull(splitManager, "splitManager is null");
-        this.functionManager = requireNonNull(functionManager, "functionManager is null");
-        this.functionResolution = requireNonNull(functionResolution, "functionResolution is null");
-    }
-
-    @Override
-    public ConnectorPlanOptimizerProvider getConnectorPlanOptimizerProvider()
-    {
-        return new ClpPlanOptimizerProvider(functionManager, functionResolution);
-    }
-
-    @Override
-    public ConnectorTransactionHandle beginTransaction(IsolationLevel isolationLevel, boolean readOnly)
-    {
-        return ClpTransactionHandle.INSTANCE;
-    }
-
-    @Override
-    public ConnectorMetadata getMetadata(ConnectorTransactionHandle transactionHandle)
-    {
-        return metadata;
-    }
-
-    @Override
-    public ConnectorSplitManager getSplitManager()
-    {
-        return splitManager;
-    }
-
-    @Override
-    public final void shutdown()
-    {
-        try {
-            lifeCycleManager.stop();
-        }
-        catch (Exception e) {
-            log.error(e, "Error shutting down connector");
-        }
-    }
-}
