--- presto-clp/src/main/java/com/yscope/presto/metadata/ClpMetadataProvider.java
@@ -1,26 +1,0 @@
-/*
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.yscope.presto.metadata;
-
-import com.facebook.presto.spi.SchemaTableName;
-import com.yscope.presto.ClpColumnHandle;
-
-import java.util.List;
-
-public interface ClpMetadataProvider
-{
-    public List<ClpColumnHandle> listColumnHandles(SchemaTableName schemaTableName);
-
-    public List<String> listTableNames(String schema);
-}
