+++ presto-server/src/main/provisio/presto.xml
@@ -231,7 +231,7 @@
     </artifactSet>
 
     <artifactSet to="plugin/presto-clp">
-        <artifact id="com.yscope.presto:presto-clp:zip:${project.version}">
+        <artifact id="${project.groupId}:presto-clickhouse:zip:${project.version}">
             <unpack />
         </artifact>
     </artifactSet>
